requirements.txt
certifi==2022.12.7
cftime==1.6.2
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
GDAL==3.6.3
kiwisolver==1.4.4
matplotlib==3.7.1
munkres==1.1.4
netCDF4==1.6.3
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pip==23.0.1
pygmt==0.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
setuptools==67.6.1
six==1.16.0
tornado==6.2
tzdata==2023.3
wheel==0.40.0
xarray==2022.11.0
